:lipstick: data.json 정보로 데이터 대체
