--- src/components/SavedStories.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const SavedStories = props => {
-    return (
-        <div>
-
-        </div>
-    );
-};
-
-export default SavedStories;
