+++ src/pages/ProfilePage.jsx
@@ -1,11 +1,11 @@
 // import React from 'react';
 
-import UserSection from "@/components/UserSection.jsx";
-import SavedStories from "@/components/SavedStories.jsx";
+import UserSection from "@/components/userSection/UserSection.jsx";
+import SavedStories from "@/components/savedStories/SavedStories.jsx";
 
 const ProfilePage = () => {
     return (
-        <div className="m-16 p-4">
+        <div className="m-16 p-4 w-[935px]">
             <UserSection/>
             <SavedStories/>
